Creating updateAccsLend.test.js and updaating updateAccLend function
